Provide default mode for Volume.mkdir()
Calling mkdir without mode used to fail with following error:

>>> vol.mkdir('/c2')
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  TypeError: mkdir() takes exactly 3 arguments (2 given)

Python's os.mkdir() defined default mode:
https://docs.python.org/2/library/os.html#os.mkdir

Also, changed mode_t is of type unsigned short (ctypes.c_ushort).

+++ glusterfs/api.py
@@ -213,7 +213,7 @@ glfs_closedir = ctypes.CFUNCTYPE(
 
 
 glfs_mkdir = ctypes.CFUNCTYPE(ctypes.c_int, ctypes.c_void_p, ctypes.c_char_p,
-                              ctypes.c_ulong)(('glfs_mkdir', client))
+                              ctypes.c_ushort)(('glfs_mkdir', client))
 
 glfs_opendir = ctypes.CFUNCTYPE(ctypes.c_void_p,
                                 ctypes.c_void_p,
